int to long

+++ parasol-tags/parasol-tags-api/src/main/java/com/ginkgocap/parasol/tags/model/TagSource.java
@@ -96,7 +96,7 @@ public class TagSource implements Serializable {
 		return sourceType;
 	}
 
-	public void setSourceType(int sourceType) {
+	public void setSourceType(long sourceType) {
 		this.sourceType = sourceType;
 	}
